% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_expression.R
\name{calculate_correct_pseudotime}
\alias{calculate_correct_pseudotime}
\title{Get corrected pseudotime}
\usage{
calculate_correct_pseudotime(dataset, start, normalized = TRUE)
}
\arguments{
\item{dataset}{A dyngen dataset}

\item{start}{The start milestone}
}
\description{
Get corrected pseudotime
}
