% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_expression.R
\name{calculate_pseudotime}
\alias{calculate_pseudotime}
\title{Get uncorrected pseudotime}
\usage{
calculate_pseudotime(dataset, milestone_network, start)
}
\arguments{
\item{dataset}{A dyngen dataset}
}
\description{
Get uncorrected pseudotime
}
