% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\title{Combine multiple dyngen models}
\usage{
combine_models(models)
}
\arguments{
\item{models}{A list of models.}
}
\description{
Assume the given models ran up to `generate_cells()`, and
will run `generate_experiment()` on the combined models.
}
\examples{
\donttest{
# create the config object
init_config <- initialise_model(
  backbone = backbone_linear(),
  num_cells = 500,
  num_targets = 250,
  num_hks = 50,
  gold_standard_params = gold_standard_default(census_interval = 1, tau = 0.05),
  simulation_params = simulation_default(
    # burn_time = 10,
    # total_time = 10,
    census_interval = 1,
    ssa_algorithm = ssa_etl(tau = 0.05),
    experiment_params = simulation_type_wild_type(num_simulations = 40)
  )
)

# generate the genes and their kinetics
model_common <-
  init_config \%>\%
  generate_tf_network() \%>\%
  generate_feature_network()
plot_feature_network(model_common)

# run the simulation once
model_a <- model_common \%>\%
  generate_kinetics() \%>\%
  generate_gold_standard() \%>\%
  generate_cells()

# run the simulation once more
model_b <- model_common \%>\%
  generate_kinetics() \%>\%
  generate_gold_standard() \%>\%
  generate_cells()

# combine models, do experiment afterwards
model_ab <- combine_models(list("left" = model_a, "right" = model_b) \%>\%
  generate_experiment()

# show a dimensionality reduction
plot_simulations(model_ab)
plot_gold_mappings(model_ab, do_facet = FALSE)

# create a dynwrap dataset
dataset <- wrap_dataset(model_ab)
}
}
