% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_expression.R
\name{compute_pseudotime_from_root}
\alias{compute_pseudotime_from_root}
\title{Get corrected pseudotime}
\usage{
compute_pseudotime_from_root(
  dataset,
  root = unique(setdiff(dataset$milestone_network$from, dataset$milestone_network$to)),
  normalized = TRUE
)
}
\arguments{
\item{dataset}{A dyngen dataset}

\item{root}{The root milestone to compute the pseudotime from.}
}
\description{
Get corrected pseudotime
}
