% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cni_auc.R
\name{evaluate_ranking_direct}
\alias{evaluate_ranking_direct}
\title{Evaluate a ranking}
\usage{
evaluate_ranking_direct(
  values,
  are_true,
  num_positive_interactions,
  num_possible_interactions,
  extend_by = 10000
)
}
\arguments{
\item{values}{A vector of importance values of predicted interactions.}

\item{are_true}{A vector denoting whether the corresponding predicted interactions are true.}

\item{num_positive_interactions}{The total number of positives.}

\item{num_possible_interactions}{The total number ranked values.}

\item{extend_by}{The number of steps with which to fill the ranking as if random, if only a part of the ranking is given.}
}
\value{
A list containing two items, the ranked evaluation and the area under the curve scores
}
\description{
Evaluate a ranking
}
