% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cells.R
\name{filter_cells}
\alias{filter_cells}
\title{Filter cells}
\usage{
filter_cells(dataset, ..., keep_dimred = FALSE)
}
\arguments{
\item{dataset}{A dyno dataset}

\item{...}{What to filter the dataset by. Can be one of the columns in `cell_info` or a column in `progressions`.}

\item{keep_dimred}{Whether to retain the dimred or not.}
}
\description{
Filter cells
}
