% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_dataset.R
\name{generate_diff_abundance}
\alias{generate_diff_abundance}
\title{Generate difference in abundance}
\usage{
generate_diff_abundance(model, froms, tos, ratios, normalise_gs = T)
}
\description{
Generate difference in abundance
}
