% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_expression.R
\name{get_cell_expression}
\alias{get_cell_expression}
\title{Get cell expression, but sorted according to pseudotime}
\usage{
get_cell_expression(dataset, milestone_network, start)
}
\arguments{
\item{dataset}{A dyngen dataset}
}
\description{
Get cell expression, but sorted according to pseudotime
}
