% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudocell_inference.R
\name{get_waypoint_expression}
\alias{get_waypoint_expression}
\title{Get gene expression of inferred waypoints}
\usage{
get_waypoint_expression(dataset, amount, ws = 0.125)
}
\arguments{
\item{dataset}{A dyngen dataset}

\item{amount}{The amount of pseudocells wanted}

\item{ws}{The windowsize of the Gaussian kernel}
}
\description{
Get gene expression of inferred waypoints
}
