% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{\%cache\%}
\alias{\%cache\%}
\title{Obtain an object from cache, if it exists}
\usage{
file \%cache\% value
}
\arguments{
\item{file}{The cache file}

\item{value}{A code block to obtain the object}
}
\description{
Obtain an object from cache, if it exists
}
\examples{
\dontrun{
obj <- "myfile.rds" \%cache\% {
  # long execution
  Sys.sleep(100)
  10
}

}
}
