% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_simil.R
\name{paired_simil}
\alias{paired_simil}
\title{Calculate paired similarity}
\usage{
paired_simil(x, y, method = "spearman", margin = 1)
}
\arguments{
\item{x}{A numeric matrix, dense or sparse.}

\item{y}{(Optional) a numeric matrix, dense or sparse, with \code{nrow(x) == nrow(y)}.}

\item{method}{Which distance method to use. Options are: \code{"cosine"}, \code{"pearson"}, \code{"spearman"}, \code{"euclidean"}, and \code{"manhattan"}.}

\item{margin}{Which margin to use for the pairwise comparison. 1 => rowwise, 2 => columnwise.}
}
\description{
Calculate paired similarity
}
\examples{
x <- matrix(runif(500), nrow = 10)
y <- x * rnorm(500, mean = 1, sd = .05)
margin <- 1; method <- "pearson"
}
