% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size.R
\name{reduce_size}
\alias{reduce_size}
\title{Helper function for reducing the file size of dynplot2 plots}
\usage{
reduce_size(g, verbose = FALSE)
}
\arguments{
\item{g}{A dynplot2 plot}

\item{verbose}{Whether to print a message}
}
\description{
Helper function for reducing the file size of dynplot2 plots
}
