% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnav_run_velocyto.R
\name{rnav_run_velocyto}
\alias{rnav_run_velocyto}
\title{Obtain velocity with velocyto}
\usage{
rnav_run_velocyto(
  spliced,
  unspliced,
  assumption = "constant_velocity",
  velocyto = reticulate::import("velocyto")
)
}
\description{
Obtain velocity with velocyto
}
