% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{start_analysis}
\alias{start_analysis}
\title{Helper function for knowing where to store files}
\usage{
start_analysis(experiment_id)
}
\arguments{
\item{experiment_id}{The identifier of the experiment}

\item{...}{File path}
}
\description{
Helper function for knowing where to store files
}
\examples{
\dontrun{
analysis <- start_analysis("usecase_network_inference")

dataset_path <- analysis$temporary_file("dataset.rds")
temporary_file(experiment_id = "usecase_network_inference", "dataset.rds")

figure_path <- analysis$result_file("figure.pdf")
result_file(experiment_id = "usecase_network_inference", "figure.pdf")

}
}
