% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_metric.R
\name{ta_abwap}
\alias{ta_abwap}
\title{ABWAP score}
\usage{
ta_abwap(pt1, pt2)
}
\arguments{
\item{pt1}{Aligned pseudotime values for dataset 1}

\item{pt2}{Aligned pseudotime values for dataset 2}
}
\description{
ABWAP score
}
