% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_expression.R
\name{calculate_correct_pseudotime}
\alias{calculate_correct_pseudotime}
\title{Get corrected pseudotime}
\usage{
calculate_correct_pseudotime(dataset, milestone_network, start, normalized = T)
}
\arguments{
\item{dataset}{A dyngen dataset}
}
\description{
Get corrected pseudotime
}
