% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_dataset.R
\name{combine_models}
\alias{combine_models}
\title{Combine two dyngen models}
\usage{
combine_models(model1, model2)
}
\arguments{
\item{model1}{the first model}

\item{model2}{the second model}
}
\description{
Combine two dyngen models
}
