% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{\%cache\%}
\alias{\%cache\%}
\title{Obtain an object from cache, if it exists}
\usage{
file \%cache\% func
}
\arguments{
\item{file}{The cache file}

\item{func}{The function to obtain the object}
}
\description{
Obtain an object from cache, if it exists
}
\examples{
\dontrun{
long_execution_function <- function() {
  Sys.sleep(100)
  10
}

obj <- "myfile.rds" \%cache\% long_execution_function
}
}
